fix: Challenge 계산 로직의 금액 누적을 Long으로 전환

+++ src/main/java/Hampouch/server/domain/challenge/service/ChallengeCalculator.java
@@ -25,7 +25,7 @@ public final class ChallengeCalculator {
     }
 
     public static int recommendedBudgetTotal(ChallengeStatus status, EndReason endReason,
-                                             int budgetTotal, int actualSpent) {
+                                             int budgetTotal, long actualSpent) {
         if (status == ChallengeStatus.SUCCESS) {
             if (actualSpent > budgetTotal) {
                 throw new IllegalStateException("성공한 챌린지의 실지출이 목표 금액을 초과했습니다.");
@@ -39,13 +39,15 @@ public final class ChallengeCalculator {
             if (actualSpent <= budgetTotal) {
                 throw new IllegalStateException("금액 초과로 실패한 챌린지의 실지출이 목표 금액 이하입니다.");
             }
-            int overAmount = actualSpent - budgetTotal;
-            return budgetTotal + overAmount / 2 + overAmount % 2;
+            long overAmount = actualSpent - budgetTotal;
+            // 새 목표도 결국 budgetTotal 자리(BUDGET_TOTAL_MAX 이내)에 들어갈 값이라 int로 좁힌다 —
+            // 넘치면 조용히 틀린 추천값을 주는 대신 예외로 드러나게 Math.toIntExact를 쓴다.
+            return Math.toIntExact(budgetTotal + overAmount / 2 + overAmount % 2);
         }
         throw new IllegalArgumentException("종료되지 않은 챌린지는 추천할 수 없습니다: " + status);
     }
 
-    public static DayStatus judge(int spentAmount, int dailyLimit) {
+    public static DayStatus judge(long spentAmount, int dailyLimit) {
         return spentAmount <= dailyLimit ? DayStatus.SUCCESS : DayStatus.OVER;
     }
 
@@ -64,9 +66,9 @@ public final class ChallengeCalculator {
 
         int successDays = 0;
         int overDays = 0;
-        int savedAmount = 0;
-        int overAmount = 0;
-        int actualSpent = 0;
+        long savedAmount = 0;
+        long overAmount = 0;
+        long actualSpent = 0;
         int maxStreak = 0;
         int currentStreak = 0;
 
@@ -98,18 +100,18 @@ public final class ChallengeCalculator {
      * summarizeThrough의 Expense 기반 버전 — 날짜별 지출 합계 맵을 받아 한도·판정을 계산
      * DailyLimitTimeline.on(date)는 조정 이력만으로 결정되는 순수 함수라 얼려둘 필요가 없어 항상 limits.on(date)를 쓴다.
      */
-    public static ChallengeSummary summarizeThrough(Map<LocalDate, Integer> spentByDate, DailyLimitTimeline limits,
+    public static ChallengeSummary summarizeThrough(Map<LocalDate, Long> spentByDate, DailyLimitTimeline limits,
                                                      LocalDate startDate, LocalDate endDate) {
         int successDays = 0;
         int overDays = 0;
-        int savedAmount = 0;
-        int overAmount = 0;
-        int actualSpent = 0;
+        long savedAmount = 0;
+        long overAmount = 0;
+        long actualSpent = 0;
         int maxStreak = 0;
         int currentStreak = 0;
 
         for (LocalDate date = startDate; !date.isAfter(endDate); date = date.plusDays(1)) {
-            int spent = spentByDate.getOrDefault(date, 0);
+            long spent = spentByDate.getOrDefault(date, 0L);
             int dailyLimit = limits.on(date);
             DayStatus status = judge(spent, dailyLimit);
             actualSpent += spent;
@@ -142,11 +144,11 @@ public final class ChallengeCalculator {
      * currentStreakAsOf의 Expense 기반 버전.
      * 이 버전은 status가 없으므로 judge(spent, limits.on(date))로 그 자리에서 계산 — limits 파라미터가 필요
      */
-    public static int currentStreakAsOf(Map<LocalDate, Integer> spentByDate, DailyLimitTimeline limits,
+    public static int currentStreakAsOf(Map<LocalDate, Long> spentByDate, DailyLimitTimeline limits,
                                         LocalDate startDate, LocalDate aggregationEndDate) {
         int streak = 0;
         for (LocalDate date = aggregationEndDate; !date.isBefore(startDate); date = date.minusDays(1)) {
-            int spent = spentByDate.getOrDefault(date, 0);
+            long spent = spentByDate.getOrDefault(date, 0L);
             if (judge(spent, limits.on(date)) != DayStatus.SUCCESS) {
                 break;
             }
@@ -156,7 +158,7 @@ public final class ChallengeCalculator {
     }
 
     /** 한도 0원일 때 Infinity·NaN이 응답에 노출되지 않도록 초과 여부만 반환한다. */
-    public static double usageRate(int todaySpent, int dailyLimit) {
+    public static double usageRate(long todaySpent, int dailyLimit) {
         if (dailyLimit <= 0) {
             return todaySpent > 0 ? 1.0 : 0.0;
         }
@@ -201,7 +203,7 @@ public final class ChallengeCalculator {
      * 일별 초과(OVER)는 달력 표시·overDays 집계로만 남고 성패를 가르지 않는다.
      * actualSpent는 summarizeThrough가 만든 값을 넘길 것 — 판정 근거와 응답의 총액이 같은 계산이어야 한다.
      */
-    public static ChallengeStatus resultStatus(int actualSpent, int budgetTotal) {
+    public static ChallengeStatus resultStatus(long actualSpent, int budgetTotal) {
         return actualSpent > budgetTotal ? ChallengeStatus.FAIL : ChallengeStatus.SUCCESS;
     }
